+++ PokemonTcgSdkv2/Client/ApiClient.cs
@@ -8,7 +8,7 @@ using System.Threading.Tasks;
 using PokemonTcgSdkV2.Api;
 using PokemonTcgSdkV2.Client.Endpoints;
 using PokemonTcgSdkV2.Client.Responses;
-using PokemonTcgSdkV2.Utils;
+using PokemonTcgSdkV2.Utils.Query;
 
 namespace PokemonTcgSdkV2.Client
 {
